hardware/requirements.txt
rclpy==3.3.2
numpy==1.24.3
pyserial==3.5
opencv-python==4.8.0.76
python-dotenv==1.0.0
paho-mqtt==1.6.1
RPi.GPIO==0.7.1
adafruit-circuitpython-motorkit==1.6.8 